requirements.txt
aiohttp==2.2.5
astroid==1.6.1
async-timeout==2.0.0
chardet==3.0.4
discord.py==1.0.0a0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.1.0
pylint==1.8.2
six==1.11.0
websockets==4.0.1
wrapt==1.10.11
yarl==1.1.1
